Struct timeseries::Table

pub struct Table<U, V> where
    U: Ord,  {
    pub headers: Vec<String>,
    pub data: BTreeMap<U, Vec<V>>,
    pub meta_data: Option<HashMap<String, String>>,
}

 use timeseries::{vec2, Table};

 let headers = vec![
     String::from("number"),
     String::from("text"),
     String::from("test"),
     String::from("data")
 ];

 let indexes = vec![1, 2, 3, 4, 5, 6];
 let data = vec2![
     ["1", "Test01", "test", "abcd"],
     ["2", "Test02", "test", "efgh"],
     ["3", "Test03", "test", "ijkl"],
     ["4", "Test04", "test", "mnop"],
     ["5", "Test05", "test", "qrst"],
     ["6", "Test06", "test", "uvwx"],
 ];

  Table::new(headers, indexes, data).unwrap();

Fields
headers: Vec<String>data: BTreeMap<U, Vec<V>>meta_data: Option<HashMap<String, String>>
Methods

impl<U, V> Table<U, V> where
U: Ord,

pub fn new_btreemap(
headers: Vec<String>,
data: BTreeMap<U, Vec<V>>
) -> Table<U, V>

pub fn new(
headers: Vec<String>,
indexes: Vec<U>,
data: Vec<Vec<V>>
) -> Result<Table<U, V>, TableError>

pub fn from_map(map: &HashMap<U, HashMap<String, V>>) -> Table<U, V> where
U: Hash + Clone,
V: Clone,

Creates table from data in this form:

 {
   1: {
        "test": "testing",
        "number": "1"
      },
   2: {
        "test": "testing",
        "number": "2"
      },
   3: {
        "test": "testing",
        "number": "3"
      }
 }

pub fn from_map_checked(
map: &HashMap<U, HashMap<String, V>>
) -> Result<Table<U, V>, TableError> where
U: Hash + Clone,
V: Clone,

Creates table from data in this form and returns an error if it does not fit:

 {
   1: {
        "test": "testing",
        "number": "1"
      },
   2: {
        "test": "testing",
        "number": "2"
      },
   3: {
        "test": "testing",
        "number": "3"
      }
 }

pub fn from_map_safe(
map: &HashMap<U, HashMap<String, V>>
) -> Table<U, Option<V>> where
U: Hash + Clone,
V: Clone,

Creates table from data in this form, and return a table where every cell is wrapped in an option object:

 {
   1: {
        "test": "testing",
        "number": "1"
      },
   2: {
        "test": "testing",
        "number": "2",
        "extra_column": "info"
      },
   3: {
        "test": "testing",
        "number": "3"
      }
 }

Important traits for IterRows<'a, K, V, H>
impl<'a, K, V, H> Iterator for IterRows<'a, K, V, H> type Item = Row<'a, K, V, H>;

pub fn iter_rows(&mut self) -> IterRows<U, V, String>